Refactor hours to separate component
