+++ src/components/SelectedTap.js
@@ -14,7 +14,6 @@ import styles from "./SelectedTap.module.scss";
 import sampleImg from "./fountain.png";
 import phlaskGreen from "./images/phlaskGreen.png";
 import phlaskBlue from "./images/phlaskBlue.png";
-import { hours } from "./hours.js";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faCaretLeft,
@@ -23,6 +22,7 @@ import {
   faTimes
 } from "@fortawesome/free-solid-svg-icons";
 import SelectedTapIcons from "./SelectedTapIcons";
+import SelectedTapHours from "./SelectedTapHours";
 
 const tempImages = {
   tapImg: sampleImg
@@ -37,16 +37,8 @@ class SelectedTap extends React.Component {
     infoExpansionStyle: {},
     isDescriptionShown: false,
     tapDescription: null,
-    isHoursExpanded: false,
     animationSpeed: 600,
-    currentDay: null,
-    currentHour: "",
-    currentMinute: "",
-    hoursList: null,
-    currentOrgHours: null,
-    isOpen: null,
     organization: this.props.selectedPlace.organization,
-    hours: this.props.selectedPlace.hours,
     address: this.props.selectedPlace.address,
     accessible: this.props.selectedPlace.accessible,
     testIcons: {
@@ -73,9 +65,6 @@ class SelectedTap extends React.Component {
       if (!this.props.infoIsExpanded) {
         this.toggleInfoWindow(false);
       }
-      this.setState({
-        isHoursExpanded: false
-      });
     } else {
       // Set height on first render to animate expansion
       if (Object.keys(this.state.infoExpansionStyle).length < 1) {
@@ -159,91 +148,16 @@ class SelectedTap extends React.Component {
   // Handle Times
 
   setCurrentDate() {
-    const today = new Date();
-    const currentDay = today.getDay();
-    // const hour = today.getHours()
-    // const getMinute = today.getMinutes()
-    // console.log('Time: ' + today.getHours().toString() + today.getMinutes().toString());
-
     const selectedPlace = this.props.selectedPlace;
 
     this.setState({
-      currentDay: currentDay,
-      currentHour: hours.getHourFromMilitary(today.getHours()),
-      currentMinute: today.getMinutes(),
-      hoursList: selectedPlace.hours !== undefined ? this.getAllHours() : null,
-      currentOrgHours:
-        selectedPlace.hours !== undefined
-          ? selectedPlace.hours[currentDay] !== undefined
-            ? selectedPlace.hours[currentDay].open !== undefined &&
-              selectedPlace.hours[currentDay].close !== undefined
-              ? {
-                  open: hours.getSimpleHours(
-                    selectedPlace.hours[currentDay].open.time
-                  ),
-                  close: hours.getSimpleHours(
-                    selectedPlace.hours[currentDay].close.time
-                  )
-                }
-              : false
-            : false
-          : null,
       organization: selectedPlace.organization,
       address: selectedPlace.address,
-      isOpen:
-        selectedPlace.hours !== undefined
-          ? selectedPlace.hours[currentDay]
-            ? selectedPlace.hours[currentDay].close !== undefined &&
-              selectedPlace.hours[currentDay].open !== undefined
-              ? hours.checkOpen(
-                  selectedPlace.hours[currentDay].open.time,
-                  selectedPlace.hours[currentDay].close.time
-                )
-              : false
-            : false
-          : null,
       tapDescription:
         selectedPlace.description !== undefined
           ? selectedPlace.description
           : null
     });
-  }
-
-  /* Return an array of objects containing Day of the week, open time, 
-        and closing time, starting with the current day
-     */
-  getAllHours() {
-    const selectedPlace = this.props.selectedPlace;
-
-    if (selectedPlace.hours === undefined) {
-      return null;
-    } else {
-      const hoursList = [];
-
-      selectedPlace.hours.map((orgHours, index) => {
-        const formattedHours = {
-          day: hours.getDays(index),
-          open:
-            orgHours.open !== undefined
-              ? hours.getSimpleHours(orgHours.open.time)
-              : null,
-          close:
-            orgHours.close !== undefined
-              ? hours.getSimpleHours(orgHours.close.time)
-              : null
-        };
-        hoursList.push(formattedHours);
-      });
-
-      // Shift array so current day is first
-      const date = new Date();
-      const day = date.getDay();
-      for (let x = 0; x < day; x++) {
-        hoursList.push(hoursList.shift());
-      }
-
-      return hoursList;
-    }
   }
 
   // getAccess(){
@@ -278,9 +192,6 @@ class SelectedTap extends React.Component {
   componentDidUpdate(prevProps) {
     if (this.props.showingInfoWindow) {
       if (this.props.selectedPlace !== prevProps.selectedPlace) {
-        if (this.props.selectedPlace.hours !== undefined) {
-          // console.log('Did Update. Props: ' + this.props.selectedPlace);
-        }
         this.setCurrentDate();
       }
       if (
@@ -422,133 +333,10 @@ class SelectedTap extends React.Component {
                 )}
               </div>
 
-              {/* Hours */}
-              <div id="org-hours">
-                <div
-                  id="tap-info-org-status"
-                  style={
-                    this.state.isOpen
-                      ? { color: "green" }
-                      : this.state.isOpen !== null
-                      ? { color: "red" }
-                      : { color: "orange" }
-                  }
-                >
-                  {this.state.isOpen
-                    ? "Open"
-                    : this.state.isOpen !== null
-                    ? "Closed"
-                    : "unavailable"}
-                </div>
-                {this.state.currentOrgHours && (
-                  <div id="hours-area">
-                    {/* Placeholder for Dropdown */}
-
-                    <div
-                      id="tap-info-hours-container-placeholder"
-                      style={
-                        this.state.currentOrgHours !== null
-                          ? { border: "1px solid #c4c4c4" }
-                          : {}
-                      }
-                    >
-                      <div className="tap-hours-list-item">Placeholder</div>
-                      {/* <div className='hours-dropdown-arrow-container' style={{width: this.props.infoIsExpanded ? '20px' : '0' }}>
-                                            <img className='hours-dropdown-arrow' src={hoursArrow} alt=''></img>
-                                        </div> */}
-                    </div>
-
-                    {/* Container of all visible hours elements */}
-                    <div
-                      id="tap-info-hours-container"
-                      style={
-                        this.state.currentOrgHours !== null
-                          ? { border: "1px solid #c4c4c4" }
-                          : {}
-                      }
-                    >
-                      {/* Placeholder for Dropdown */}
-                      <div id="current-hours-placeholder">
-                        <div className="tap-hours-list-item">
-                          {this.state.currentOrgHours !== null
-                            ? this.state.currentOrgHours !== false
-                              ? `${this.state.currentOrgHours.open} - ${this.state.currentOrgHours.close}`
-                              : ""
-                            : ""}
-                        </div>
-                        {(this.props.infoIsExpanded || !isMobile) &&
-                          this.state.hoursList !== null &&
-                          this.state.currentOrgHours !== false && (
-                            <div className={styles.hoursDropdownArrowContainer}>
-                              <FontAwesomeIcon
-                                className={styles.hoursDropdownArrow}
-                                color="#999"
-                                size="2x"
-                                icon={faCaretDown}
-                              />
-                            </div>
-                          )}
-                      </div>
-
-                      {/* Current Day Hours */}
-                      <div
-                        id="current-hours"
-                        onClick={() => {
-                          if (this.props.infoIsExpanded || !isMobile) {
-                            this.setState({
-                              isHoursExpanded: !this.state.isHoursExpanded
-                            });
-                          }
-                        }}
-                      >
-                        <div className="tap-hours-list-item">
-                          {this.state.currentOrgHours !== null
-                            ? this.state.currentOrgHours !== false
-                              ? `${this.state.currentOrgHours.open} - ${this.state.currentOrgHours.close}`
-                              : ""
-                            : ""}
-                        </div>
-                        {(this.props.infoIsExpanded || !isMobile) &&
-                          this.state.hoursList !== null &&
-                          this.state.currentOrgHours !== false && (
-                            <div className={styles.hoursDropdownArrowContainer}>
-                              <FontAwesomeIcon
-                                className={styles.hoursDropdownArrow}
-                                color="#999"
-                                size="2x"
-                                icon={faCaretDown}
-                              />
-                            </div>
-                          )}
-                      </div>
-                      {/* Other Days */}
-                      {this.state.isHoursExpanded &&
-                      (this.props.infoIsExpanded || !isMobile) ? (
-                        <div id="other-hours-container">
-                          {this.state.hoursList !== null ? (
-                            this.state.hoursList.map((hours, index) => {
-                              if (index !== 0) {
-                                return (
-                                  <div
-                                    className="tap-hours-list-item"
-                                    key={index}
-                                  >
-                                    {`${hours.day} ${hours.open} - ${hours.close}`}
-                                  </div>
-                                );
-                              }
-                            })
-                          ) : (
-                            <div className="tap-hours-list-item">n/a</div>
-                          )}
-                        </div>
-                      ) : (
-                        <div></div>
-                      )}
-                    </div>
-                  </div>
-                )}
-              </div>
+              <SelectedTapHours
+                infoIsExpanded={this.props.infoIsExpanded}
+                selectedPlace={this.props.selectedPlace}
+              />
             </div>
             {/* Walk Time & Info Icons */}
             <div className={styles.walkTime}>Estimated Walk Time: 12 mins</div>
