+++ src/components/SelectedTap.module.scss
@@ -31,15 +31,6 @@
   font-weight: 500;
 }
 
-.hoursDropdownArrowContainer {
-  transition: 0.6s width ease-in-out;
-}
-
-.hoursDropdownArrow {
-  width: 100%;
-  height: 1rem;
-}
-
 .description {
   padding: 0.5rem 1rem;
 }
