+++ src/components/SelectedTapHours.module.scss
@@ -1,0 +1,8 @@
+.hoursDropdownArrowContainer {
+  transition: 0.6s width ease-in-out;
+}
+
+.hoursDropdownArrow {
+  width: 100%;
+  height: 1rem;
+}
